jOOQ and Webpack part 2: migration to jOOQ and Spring boot 1.3

+++ src/main/java/org/ng200/openolympus/controller/archive/ArchiveUserRankListRestController.java
@@ -61,7 +61,7 @@ public class ArchiveUserRankListRestController {
 			+ SecurityExpressionConstants.NO_CONTEST_CURRENTLY + ')')
 	@RequestMapping(value = "/api/archive/rank", method = RequestMethod.GET)
 	@JsonView(UnprivilegedView.class)
-	public List<UserRanking> getUsers(@RequestParam("page") Long page) {
+	public List<UserRanking> getUsers(@RequestParam("page") Integer page) {
 		return this.userService.getArchiveRankPage(page,
 				ArchiveUserRankListRestController.PAGE_SIZE);
 	}
